Rework doCategorySearch
Category search is now actually using the category property, cleaner
approach for global search. Furthermore, the number of search keys has
been reduced, because we do not search for all DbRowID of all member
cards, but for unique category string, which are less or equal to the
number of unique DbRowIDs.