Rework modifyCard and bulk edit to fix issue #3
If bulk edit is used on a malinglist category, the parent book is
searched for missing contacts. If contacts are still missing, they get
added to the parent book and to the mailing list. If a contact is found
in the parent book, it just gets added to the category and to the
mailing list.