Use Map() for contacts.
In compose.mjs I forced both methods for compose_action and browser_action to use the same code, to simplify and unify the code.

Since we now always store the entire contact in the contacts Map() (even for categories), we can use it directly when iterating over contacts, instead of looking it up first.